backend/requirements.txt
pandas
requests
pyomo
flask
psycopg2-binary
python-dotenv
flask_sqlalchemy
flask_migrate
flask-login
flask_wtf
Werkzeug
email_validator